# openapi.json
{"openapi":"3.0.2","info":{"title":"Campus Facility Reservation API","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/facilities/":{"get":{"tags":["facilities"],"summary":"List Facilities","operationId":"list_facilities_facilities__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response List Facilities Facilities  Get","type":"array","items":{"$ref":"#/components/schemas/FacilityOut"}}}}}}},"post":{"tags":["facilities"],"summary":"Create Facility","operationId":"create_facility_facilities__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FacilityCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FacilityOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/facilities/{facility_id}":{"put":{"tags":["facilities"],"summary":"Update Facility","operationId":"update_facility_facilities__facility_id__put","parameters":[{"required":true,"schema":{"title":"Facility Id","type":"integer"},"name":"facility_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FacilityUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FacilityOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]},"delete":{"tags":["facilities"],"summary":"Delete Facility","operationId":"delete_facility_facilities__facility_id__delete","parameters":[{"required":true,"schema":{"title":"Facility Id","type":"integer"},"name":"facility_id","in":"path"}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/reservations/":{"get":{"tags":["reservations"],"summary":"List All Reservations","operationId":"list_all_reservations_reservations__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response List All Reservations Reservations  Get","type":"array","items":{"$ref":"#/components/schemas/ReservationOut"}}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["reservations"],"summary":"Create Reservation","operationId":"create_reservation_reservations__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/reservations/me":{"get":{"tags":["reservations"],"summary":"List My Reservations","operationId":"list_my_reservations_reservations_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response List My Reservations Reservations Me Get","type":"array","items":{"$ref":"#/components/schemas/ReservationOut"}}}}}},"security":[{"HTTPBearer":[]}]}},"/reservations/{reservation_id}/status":{"put":{"tags":["reservations"],"summary":"Update Reservation Status","operationId":"update_reservation_status_reservations__reservation_id__status_put","parameters":[{"required":true,"schema":{"title":"Reservation Id","type":"integer"},"name":"reservation_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationStatusUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"FacilityCreate":{"title":"FacilityCreate","required":["name"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string"},"location":{"title":"Location","type":"string"},"capacity":{"title":"Capacity","type":"integer"}}},"FacilityOut":{"title":"FacilityOut","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string"},"location":{"title":"Location","type":"string"},"capacity":{"title":"Capacity","type":"integer"}}},"FacilityUpdate":{"title":"FacilityUpdate","type":"object","properties":{"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string"},"location":{"title":"Location","type":"string"},"capacity":{"title":"Capacity","type":"integer"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ReservationCreate":{"title":"ReservationCreate","required":["facility_id","start_date","end_date"],"type":"object","properties":{"facility_id":{"title":"Facility Id","type":"integer"},"start_date":{"title":"Start Date","type":"string","format":"date"},"end_date":{"title":"End Date","type":"string","format":"date"}}},"ReservationOut":{"title":"ReservationOut","required":["id","user_id","facility_id","start_date","end_date","status","created_at"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"user_id":{"title":"User Id","type":"string"},"facility_id":{"title":"Facility Id","type":"integer"},"start_date":{"title":"Start Date","type":"string","format":"date"},"end_date":{"title":"End Date","type":"string","format":"date"},"status":{"title":"Status","type":"string"},"created_at":{"title":"Created At","type":"string","format":"date-time"}}},"ReservationStatusUpdate":{"title":"ReservationStatusUpdate","required":["status"],"type":"object","properties":{"status":{"title":"Status","pattern":"^(APPROVED|REJECTED)$","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}